Extract Android review content presentation parser
